src/main/java/org/jenkinsci/plugins/gravatar/model/GravatarUrlCreator.java
package org.jenkinsci.plugins.gravatar.model;

import com.google.common.annotations.VisibleForTesting;
import de.bripkens.gravatar.DefaultImage;
import de.bripkens.gravatar.Gravatar;
import jenkins.model.Jenkins;
import org.jenkinsci.plugins.gravatar.factory.GravatarFactory;

import static com.google.common.base.Preconditions.checkArgument;
import static com.google.common.base.Preconditions.checkNotNull;

public class GravatarUrlCreator {

	private final String email;

	private GravatarUrlCreator(String email) {
		this.email = email;
	}

	public static GravatarUrlCreator of(GravatarUser user) {
		checkNotNull(user);
		checkArgument(user.emailAddress().isPresent(), "Only users with e-mail address are supported");
		return new GravatarUrlCreator(user.emailAddress().get());
	}

	public String buildUrlForSize(int size) {
		checkArgument(size > 0, "Only positive sizes are allowed.");
		return gravatar().setSize(size).getUrl(email);
	}

	@VisibleForTesting
	Gravatar gravatar() {
		return new GravatarFactory().userGravatar();
	}

}
